Renamed xsd namespace from mar2016 to nov2015, to be consistent with the one in gateway 4.x
