+++ server/src/main/java/org/kaazing/gateway/server/config/parse/GatewayConfigParser.java
@@ -62,13 +62,13 @@ import org.jdom.output.XMLOutputter;
 import org.kaazing.gateway.server.Launcher;
 import org.kaazing.gateway.server.config.parse.translate.GatewayConfigTranslator;
 import org.kaazing.gateway.server.config.parse.translate.GatewayConfigTranslatorFactory;
-import org.kaazing.gateway.server.config.mar2016.ClusterType;
-import org.kaazing.gateway.server.config.mar2016.GatewayConfigDocument;
-import org.kaazing.gateway.server.config.mar2016.PropertiesType;
-import org.kaazing.gateway.server.config.mar2016.PropertyType;
-import org.kaazing.gateway.server.config.mar2016.SecurityType;
-import org.kaazing.gateway.server.config.mar2016.ServiceDefaultsType;
-import org.kaazing.gateway.server.config.mar2016.ServiceType;
+import org.kaazing.gateway.server.config.nov2015.ClusterType;
+import org.kaazing.gateway.server.config.nov2015.GatewayConfigDocument;
+import org.kaazing.gateway.server.config.nov2015.PropertiesType;
+import org.kaazing.gateway.server.config.nov2015.PropertyType;
+import org.kaazing.gateway.server.config.nov2015.SecurityType;
+import org.kaazing.gateway.server.config.nov2015.ServiceDefaultsType;
+import org.kaazing.gateway.server.config.nov2015.ServiceType;
 import org.kaazing.gateway.util.parse.ConfigParameter;
 import org.slf4j.Logger;
 import org.xml.sax.Attributes;
