+++ server/src/main/java/org/kaazing/gateway/server/context/resolve/DefaultAcceptOptionsContext.java
@@ -45,7 +45,7 @@ import java.util.Map.Entry;
 import java.util.concurrent.TimeUnit;
 
 import org.kaazing.gateway.resource.address.uri.URIUtils;
-import org.kaazing.gateway.server.config.mar2016.ServiceAcceptOptionsType;
+import org.kaazing.gateway.server.config.nov2015.ServiceAcceptOptionsType;
 import org.kaazing.gateway.service.AcceptOptionsContext;
 import org.kaazing.gateway.util.Utils;
 import org.kaazing.gateway.util.ssl.SslCipherSuites;
