+++ server/src/main/java/org/kaazing/gateway/server/context/resolve/GatewayContextResolver.java
@@ -64,22 +64,22 @@ import org.kaazing.gateway.server.Gateway;
 import org.kaazing.gateway.server.Launcher;
 import org.kaazing.gateway.server.config.SchemeConfig;
 import org.kaazing.gateway.server.config.parse.DefaultSchemeConfig;
-import org.kaazing.gateway.server.config.mar2016.AuthenticationType;
-import org.kaazing.gateway.server.config.mar2016.AuthorizationConstraintType;
-import org.kaazing.gateway.server.config.mar2016.ClusterConnectOptionsType;
-import org.kaazing.gateway.server.config.mar2016.ClusterType;
-import org.kaazing.gateway.server.config.mar2016.CrossSiteConstraintType;
-import org.kaazing.gateway.server.config.mar2016.GatewayConfigDocument;
-import org.kaazing.gateway.server.config.mar2016.LoginModuleOptionsType;
-import org.kaazing.gateway.server.config.mar2016.LoginModuleType;
-import org.kaazing.gateway.server.config.mar2016.MimeMappingType;
-import org.kaazing.gateway.server.config.mar2016.RealmType;
-import org.kaazing.gateway.server.config.mar2016.SecurityType;
-import org.kaazing.gateway.server.config.mar2016.ServiceAcceptOptionsType;
-import org.kaazing.gateway.server.config.mar2016.ServiceConnectOptionsType;
-import org.kaazing.gateway.server.config.mar2016.ServiceDefaultsType;
-import org.kaazing.gateway.server.config.mar2016.ServicePropertiesType;
-import org.kaazing.gateway.server.config.mar2016.ServiceType;
+import org.kaazing.gateway.server.config.nov2015.AuthenticationType;
+import org.kaazing.gateway.server.config.nov2015.AuthorizationConstraintType;
+import org.kaazing.gateway.server.config.nov2015.ClusterConnectOptionsType;
+import org.kaazing.gateway.server.config.nov2015.ClusterType;
+import org.kaazing.gateway.server.config.nov2015.CrossSiteConstraintType;
+import org.kaazing.gateway.server.config.nov2015.GatewayConfigDocument;
+import org.kaazing.gateway.server.config.nov2015.LoginModuleOptionsType;
+import org.kaazing.gateway.server.config.nov2015.LoginModuleType;
+import org.kaazing.gateway.server.config.nov2015.MimeMappingType;
+import org.kaazing.gateway.server.config.nov2015.RealmType;
+import org.kaazing.gateway.server.config.nov2015.SecurityType;
+import org.kaazing.gateway.server.config.nov2015.ServiceAcceptOptionsType;
+import org.kaazing.gateway.server.config.nov2015.ServiceConnectOptionsType;
+import org.kaazing.gateway.server.config.nov2015.ServiceDefaultsType;
+import org.kaazing.gateway.server.config.nov2015.ServicePropertiesType;
+import org.kaazing.gateway.server.config.nov2015.ServiceType;
 import org.kaazing.gateway.server.context.DependencyContext;
 import org.kaazing.gateway.server.context.GatewayContext;
 import org.kaazing.gateway.server.service.ServiceRegistry;
