+++ server/src/test/java/org/kaazing/gateway/server/config/parse/translate/GatewayConfigTranslatorFactoryTest.java
@@ -27,7 +27,7 @@ public class GatewayConfigTranslatorFactoryTest {
         try {
             GatewayConfigTranslatorPipeline  translator = (GatewayConfigTranslatorPipeline) GatewayConfigTranslatorFactory.newInstance().getTranslator(
                     GatewayConfigNamespace.CURRENT_NS);
-             assertTrue(translator.getTranslators().get(0) instanceof March2016Validator);
+             assertTrue(translator.getTranslators().get(0) instanceof November2015Validator);
         } catch (Exception e) {
 
             e.printStackTrace();
