+++ server/src/test/java/org/kaazing/gateway/server/context/resolve/ConnectOptionsTest.java
@@ -34,8 +34,8 @@ import org.junit.BeforeClass;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.kaazing.gateway.server.config.parse.GatewayConfigParser;
-import org.kaazing.gateway.server.config.mar2016.GatewayConfigDocument;
-import org.kaazing.gateway.server.config.mar2016.ServiceConnectOptionsType;
+import org.kaazing.gateway.server.config.nov2015.GatewayConfigDocument;
+import org.kaazing.gateway.server.config.nov2015.ServiceConnectOptionsType;
 import org.kaazing.gateway.service.ConnectOptionsContext;
 import org.kaazing.gateway.service.TransportOptionNames;
