+++ service/http.directory/src/test/scripts/org/kaazing/gateway/service/http/directory/get.index.check.whole.response.rpt
@@ -27,12 +27,10 @@ write notify BARRIER
 #We need to wait for all the data to be written first
 read await BARRIER
 read "HTTP/1.1 200 OK\r\n"
-read "Cache-Control: max-age=0\r\n"
 read "Content-Length: 119\r\n"
 read "Content-Type: text/html\r\n"
 read /Date: .*/ "\r\n"
 read /ETag: .*/ "\r\n"
-read /Expires: .*/ "\r\n"
 read /Last-Modified: .*/ "\r\n"
 read "Server: Kaazing Gateway\r\n"
 read "\r\n"
