+++ service/http.directory/src/test/scripts/org/kaazing/gateway/service/http/directory/no.server.header.rpt
@@ -23,12 +23,10 @@ write "User-Agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10.6; rv:8.0) Gecko/20
 write "\r\n"
 
 read "HTTP/1.1 200 OK\r\n"
-read "Cache-Control: max-age=0\r\n"
 read "Content-Length: 119\r\n"
 read "Content-Type: text/html\r\n"
 read /Date: .*/ "\r\n"
 read /ETag: .*/ "\r\n"
-read /Expires: .*/ "\r\n"
 read /Last-Modified: .*/ "\r\n"
 # Shouldn't have the following header
 # Server: Kaazing Gateway
