+++ service/http.directory/src/test/scripts/org/kaazing/gateway/service/http/directory/request1.request2.response1.response2.rpt
@@ -55,11 +55,9 @@ write "\r\n"
 
 # Response 1
 read "HTTP/1.1 200 OK\r\n"
-read "Cache-Control: max-age=0\r\n"
 read "Content-Length: 2777\r\n"
 read /Date: .*\r\n/
 read /ETag: .*\r\n/
-read /Expires: .*\r\n/
 read /Last-Modified: .*\r\n/
 read "Server: Kaazing Gateway\r\n"
 read "\r\n"
