+++ service/http.directory/src/test/scripts/org/kaazing/gateway/service/http/directory/uri.hex.encoded.rpt
@@ -24,12 +24,10 @@ write notify BARRIER
 
 write await BARRIER
 read "HTTP/1.1 200 OK\r\n"
-read "Cache-Control: max-age=0\r\n"
 read "Content-Length: 119\r\n"
 read "Content-Type: text/html\r\n"
 read "Date: " /.*/ "\r\n"
 read /ETag: .*/ "\r\n"
-read "Expires: " /.*/ "\r\n"
 read "Last-Modified: " /.*/ "\r\n"
 read "Server: Kaazing Gateway\r\n"
 read /.*/ "\r\n"
